Minor fix for #2710
