+++ public/less/_repository.less
@@ -553,6 +553,9 @@
 					margin-right: 30px;
 				}
 			}
+			.ui.label {
+				font-size: 1em;
+			}
 		}
 	}
