+++ templates/.VERSION
@@ -1,1 +1,1 @@
-0.8.45.0225
+0.8.46.0227
